Add methods to make use of GPU local texture data copying bypassing a roundtrip to the CPU and back, fixed sf::Font::cleanup not shrinking its allocated pixel buffer storage when the user loads a new font using the same sf::Font object.

+++ include/SFML/Graphics/Font.hpp
@@ -346,16 +346,16 @@ private:
     ////////////////////////////////////////////////////////////
     // Member data
     ////////////////////////////////////////////////////////////
-    void*                      m_library;     ///< Pointer to the internal library interface (it is typeless to avoid exposing implementation details)
-    void*                      m_face;        ///< Pointer to the internal font face (it is typeless to avoid exposing implementation details)
-    void*                      m_streamRec;   ///< Pointer to the stream rec instance (it is typeless to avoid exposing implementation details)
-    void*                      m_stroker;     ///< Pointer to the stroker (it is typeless to avoid exposing implementation details)
-    int*                       m_refCount;    ///< Reference counter used by implicit sharing
-    Info                       m_info;        ///< Information about the font
-    mutable PageTable          m_pages;       ///< Table containing the glyphs pages by character size
-    mutable std::vector<Uint8> m_pixelBuffer; ///< Pixel buffer holding a glyph's pixels before being written to the texture
+    void*                       m_library;     ///< Pointer to the internal library interface (it is typeless to avoid exposing implementation details)
+    void*                       m_face;        ///< Pointer to the internal font face (it is typeless to avoid exposing implementation details)
+    void*                       m_streamRec;   ///< Pointer to the stream rec instance (it is typeless to avoid exposing implementation details)
+    void*                       m_stroker;     ///< Pointer to the stroker (it is typeless to avoid exposing implementation details)
+    int*                        m_refCount;    ///< Reference counter used by implicit sharing
+    Info                        m_info;        ///< Information about the font
+    mutable PageTable           m_pages;       ///< Table containing the glyphs pages by character size
+    mutable std::vector<Uint32> m_pixelBuffer; ///< Pixel buffer holding a glyph's pixels before being written to the texture
     #ifdef SFML_SYSTEM_ANDROID
-    void*                      m_stream; ///< Asset file streamer (if loaded from file)
+    void*                       m_stream;      ///< Asset file streamer (if loaded from file)
     #endif
 };
